Cleanup JsonParserGenerator, refactor
