+++ codegen/src/test/kotlin/software/amazon/smithy/rust/codegen/smithy/generators/HttpProtocolTestGeneratorTest.kt
@@ -144,7 +144,7 @@ class HttpProtocolTestGeneratorTest {
                 )
             }
 
-            override fun fromResponseImpl(implBlockWriter: RustWriter, operationShape: OperationShape) {
+            override fun operationImplBlock(implBlockWriter: RustWriter, operationShape: OperationShape) {
                 fromResponseFun(implBlockWriter, operationShape) {
                     writeWithNoFormatting(correctResponse)
                 }
